Restructure repository tree.
This repository was originally conceived as a TicTacToe application
whose primary purpose is as sample code used in the "Testing Strategies
for Modern Perl" ebook. However, we may want to use other sample
projects and sample files at other points in the ebook, and we want to
share all the sample files together.

This change restructures the repository tree to put all the TicTacToe
application files under a TicTacToe subdirectory.